clean

+++ proccess.py
@@ -2,7 +2,6 @@ import subprocess
 import os
 import time
 import logging
-import shlex
 
 log = logging
 log.basicConfig(filename = '/tmp/taskmaster.log', level=logging.DEBUG)
@@ -13,21 +12,39 @@ class Proccess:
 		self.name = name
 		self.command = data["command"]
 		if "stdout" in data:
-			f = open(data["stdout"],"a+")
-			self.stdout = f
+			self.stdout = data["stdout"]
 		else:
 			self.stdout = None
 		if "stderr" in data:
-			f = open(data["stderr"],"a+")
-			self.stderr = f
+			self.stderr = data["stderr"]
 		else:
 			self.stderr = None
 		self.proccess = None
 		self.pid = None
 		self.statuss = "STOPPED"
 		self.starttime = None
+		self.rc = None
+		self.startednb = 0
+		if "umask" in data:
+			self.umask = data["umask"]
+		else:
+			self.umask = 022
+		if "returncodes" in data:
+			self.returncodes = data["returncodes"]
+		else:
+			self.returncodes = None
+		if "signal" in data:
+			self.signal = data["signal"]
+		else:
+			self.signal = None
+		if "running" in data:
+			self.running = data["running"]
+		else:
+			self.running = 0
 		if "restart" in data:
 			self.restart = data["restart"]
+		else:
+			self.restart = None
 		if "expected" in data:
 			self.expected = data["expected"]
 		if "exitcodes" in data:
@@ -40,28 +57,50 @@ class Proccess:
 			self.signal = data["signal"]
 		if "timetoexit" in data:
 			self.timetoexit = data["timetoexit"]
+		if "workingdir" in data:
+			self.workingdir = data["workingdir"]
+		else:
+			self.workingdir = None
 		if "autostart" in data:
-			self.start(init=1)
+			self.start()
 
 
 
 	def start(self, init=0):
 		if init == 0:
 			print "Started : "+self.name
-		self.proccess = subprocess.Popen(
-			self.command,
-			shell = True,
-			env = os.environ,
-			stdin = subprocess.PIPE,
-			stdout = self.stdout,
-			stderr = self.stderr);
-		self.pid = self.proccess.pid
-		self.statuss = "RUNNING"
-		self.starttime = time.time()
-		log.info(" started : "+self.name + " uptime : "+time.strftime("%H:%M:%S", time.gmtime(self.starttime)))
+		try:
+			oldmask = os.umask(self.umask)
+			if self.stdout != None:
+				stdout = open(self.stdout, "a+")
+			else:
+				stderr = None
+			if self.stderr != None:
+				stderr = open(self.stderr, "a+")
+			else:
+				stderr = None
+			self.proccess = subprocess.Popen(
+				self.command,
+				shell = True,
+				env = os.environ,
+				stdin = subprocess.PIPE,
+				stdout = stdout,
+				stderr = stderr);
+			os.umask(oldmask)
+			self.pid = self.proccess.pid
+			self.startednb += 1
+			if (self.running > 0):
+				self.statuss = "STARTING"
+			else:
+				self.statuss = "RUNNING"
+			self.starttime = time.time()
+			log.info(" started : "+self.name + " uptime : "+time.strftime("%H:%M:%S", time.gmtime(self.starttime)))
+		except Exception, e:
+			log.warning("failed to launch"+self.name + ': ' + str(e))
+			print "failed to launch "+self.name+ " error : "+ str(e)
 
 	def status(self):
-		if (self.statuss == "RUNNING"):
+		if (self.statuss == "RUNNING" or self.statuss == "STARTING"):
 			timer = time.time()
 			time_delta = time.gmtime(timer - self.starttime)
 			curr_time = time.strftime("%H:%M:%S", time_delta)
@@ -72,10 +111,28 @@ class Proccess:
 	def stop(self):
 		print "Stopped : "+self.name
 		log.info(" stopped : "+ self.name)
-		self.proccess.terminate()
+		if self.signal:
+			try:
+				self.proccess.seng_signal(self.signal)
+			except Exception, e:
+				log.error(self.name+" signal not permitted")
+				print self.name +" signal not working"
+		else:
+			self.proccess.terminate()
 		self.pid = None
 		self.statuss = "STOPPED"
 		self.starttime = None
 
 	def check(self):
-		print "check"
+		if self.proccess.poll() != None:
+			self.rc = self.proccess.poll()
+			self.pid = None
+			self.statuss = "STOPPED"
+			self.starttime = None
+			if self.restart == "always" or self.restartnb >= self.startednb:
+				self.start()
+			if self.restart == "unexpected" and self.rc not in self.returncodes:
+				self.start()
+		elif self.running > 0:
+			if time.time() - self.starttime >= self.running:
+				self.statuss = "RUNNING"
